refactor(budget/me): jauge user-facing — passer du badge header à une carte sous la barre de saisie
- Renommage BudgetBadge.tsx → BudgetCard.tsx (git mv, historique préservé)
- Nouveau rendu : carte horizontale avec titre + barre de progression colorée + texte chiffré explicite ("$4.20 consommés sur $10.00 ce mois")
- Placement : sous <ChatForm /> dans ChatView.tsx, permanent (landing ET conversation), au-dessus de SuggestionGrid quand il est affiché
- Retiré du Header.tsx (import + usage)
- Réutilise budgetColor (variantes bar + text) et formatUSD depuis credits.ts
- 2 nouvelles clés EN (com_budget_card_title, com_budget_card_consumption), com_budget_badge_tooltip existante conservée pour le tooltip
- V1 informatif uniquement, garde-fous d'affichage inchangés
