+++ client/src/locales/en/translation.json
@@ -227,6 +227,8 @@
   "com_budget_action_edit": "Edit",
   "com_budget_alert_over_threshold": "{{count}} user(s) exceed 80% of their threshold",
   "com_budget_badge_tooltip": "{{percent}}% of your monthly budget — {{spent}} of {{budget}} this month",
+  "com_budget_card_consumption": "{{spent}} spent of {{budget}} this month",
+  "com_budget_card_title": "My budget this month",
   "com_budget_col_action": "Action",
   "com_budget_col_progress": "Progress",
   "com_budget_col_spent": "Consumed",
